some styles & small refactor
